@@ -9,7 +9,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"testing"
 
 	"github.com/cosnicolaou/pbzip2"
@@ -21,7 +20,7 @@ func TestHandlingFalsePositives(t *testing.T) {
 	filename := bzip2Files["300KB1"]
 
 	rd := openBzipFile(t, filename)
-	origData, err := ioutil.ReadAll(rd)
+	origData, err := io.ReadAll(rd)
 	if err != nil {
 		t.Fatal(err)
 	}
