@@ -10,7 +10,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -180,7 +179,7 @@ func TestCancelation(t *testing.T) {
 	ctx, cancel := context.WithCancel(ctx)
 	drd := pbzip2.NewReader(ctx, rd)
 	cancel()
-	_, err := ioutil.ReadAll(drd)
+	_, err := io.ReadAll(drd)
 	if err == nil || err.Error() != "context canceled" {
 		t.Errorf("expected an error or different error to the one received: %v", err)
 	}
@@ -191,7 +190,7 @@ func TestReaderErrors(t *testing.T) {
 	ctx := context.Background()
 	rd := bytes.NewBuffer(nil)
 	drd := pbzip2.NewReader(ctx, rd)
-	_, err := ioutil.ReadAll(drd)
+	_, err := io.ReadAll(drd)
 	if err == nil || err.Error() != "failed to read stream header: EOF" {
 		t.Errorf("expected an error or different error to the one received: %v", err)
 	}
@@ -199,7 +198,7 @@ func TestReaderErrors(t *testing.T) {
 	testError := func(buf []byte, msg string) {
 		rd := bytes.NewBuffer(buf)
 		drd := pbzip2.NewReader(ctx, rd)
-		_, err = ioutil.ReadAll(drd)
+		_, err = io.ReadAll(drd)
 		if err == nil || !strings.Contains(err.Error(), msg) {
 			_, _, line, _ := runtime.Caller(1)
 			t.Errorf("line: %v expected an error or different error to the one received: %v", line, err)
@@ -207,7 +206,7 @@ func TestReaderErrors(t *testing.T) {
 	}
 
 	drd = pbzip2.NewReader(ctx, &errorReader{})
-	_, err = ioutil.ReadAll(drd)
+	_, err = io.ReadAll(drd)
 	if err == nil || !strings.Contains(err.Error(), "failed to read stream header: oops") {
 		t.Errorf("expected an error or different error to the one received: %v", err)
 	}
